Moved test fixtures outside of tests folder (all files are auto included by testing suite)
